Skilt ut i egen fil det som handler om inntektsrader. Dro også ut types siden det var 3 interfaces som nå ble brukt begge stedene.
