+++ src/sider/ftrl/saksbehandling/initialStegArray.tsx
@@ -4,7 +4,7 @@ import { VurderingVirksomhet } from "./stegKomponenter/vurderingVirksomhet";
 import { VurderingPerioder } from "./stegKomponenter/vurderingPeriode/vurderingPerioder";
 import { VurderingBestemmelse } from "./stegKomponenter/vurderingBestemmelse/vurderingBestemmelse";
 import { VurderingTrygdeavgiftGammel } from "./stegKomponenter/vurderingTrygdeavgiftGammel";
-import { VurderingTrygdeavgift } from "./stegKomponenter/vurderingTrygdeavgift";
+import { VurderingTrygdeavgift } from "./stegKomponenter/vurderingTrygdeavgift/vurderingTrygdeavgift";
 import { VurderingVedtak } from "./stegKomponenter/vurderingVedtak";
 
 const initialInngangSteg = {
