--- src/sider/ftrl/saksbehandling/stegKomponenter/vurderingTrygdeavgift/index.ts
@@ -1,3 +1,0 @@
-import { VurderingTrygdeavgift } from "./vurderingTrygdeavgift";
-
-export { VurderingTrygdeavgift };
