+++ src/sider/ftrl/saksbehandling/stegKomponenter/vurderingTrygdeavgift/types.ts
@@ -1,0 +1,15 @@
+import { FieldValues } from "react-hook-form";
+
+export interface Inntekstsrad {
+  inntektskilde?: string;
+  arbAvgBetales?: string;
+  bruttoInntekt?: string;
+}
+
+export interface FieldArrayProps {
+  inntektsrader: Inntekstsrad[];
+}
+
+export type FormValuesProps = FieldValues & {
+  skattepliktig?: string;
+} & FieldArrayProps;
